Add PatternDetector interface and move pattern detection logic away from Schaapi class
