+++ modules/pipeline/prefix-span-pattern-detector/src/test/kotlin/org/cafejojo/schaapi/patterndetector/prefixspan/FrequentSequenceFinderHelper.kt
