+++ modules/pipeline/prefix-span-pattern-detector/src/test/kotlin/org/cafejojo/schaapi/patterndetector/prefixspan/FrequentSequenceFinderTest.kt
@@ -1,18 +1,18 @@
 package org.cafejojo.schaapi.patterndetector.prefixspan
 
 import org.assertj.core.api.Assertions.assertThat
-import org.cafejojo.schaapi.patterndetector.prefixspan.PatternDetector.Companion.extractSuffixes
+import org.cafejojo.schaapi.patterndetector.prefixspan.FrequentSequenceFinder.Companion.extractSuffixes
 import org.jetbrains.spek.api.Spek
 import org.jetbrains.spek.api.dsl.describe
 import org.jetbrains.spek.api.dsl.it
 
-internal class PatternDetectorTest : Spek({
+internal class FrequentSequenceFinderTest : Spek({
     describe("when looking for common sequences of simple nodes in a path") {
         it("it should find a sequence in a path of length 1") {
             val node1 = TestNode()
             val path = listOf(node1)
 
-            val detector = PatternDetector(listOf(path), 1, TestNodeComparator())
+            val detector = FrequentSequenceFinder(listOf(path), 1, TestNodeComparator())
             detector.findFrequentSequences()
 
             assertThat(detector.pathContainsSequence(path, listOf(node1))).isTrue()
@@ -23,7 +23,7 @@ internal class PatternDetectorTest : Spek({
             val node2 = TestNode()
             val path = listOf(node1)
 
-            val detector = PatternDetector(listOf(path), 1, TestNodeComparator())
+            val detector = FrequentSequenceFinder(listOf(path), 1, TestNodeComparator())
             detector.findFrequentSequences()
 
             assertThat(detector.pathContainsSequence(path, listOf(node2))).isFalse()
@@ -38,7 +38,7 @@ internal class PatternDetectorTest : Spek({
             val node6 = TestNode()
             val path = listOf(node1, node2, node3, node4, node5, node6)
 
-            val detector = PatternDetector(listOf(path), 1, TestNodeComparator())
+            val detector = FrequentSequenceFinder(listOf(path), 1, TestNodeComparator())
             detector.findFrequentSequences()
 
             assertThat(detector.pathContainsSequence(path, listOf(node1, node2))).isTrue()
@@ -53,7 +53,7 @@ internal class PatternDetectorTest : Spek({
             val node6 = TestNode()
             val path = listOf(node1, node2, node3, node4, node5, node6)
 
-            val detector = PatternDetector(listOf(path), 1, TestNodeComparator())
+            val detector = FrequentSequenceFinder(listOf(path), 1, TestNodeComparator())
             detector.findFrequentSequences()
 
             assertThat(detector.pathContainsSequence(path, listOf(node3, node4, node5))).isTrue()
@@ -68,7 +68,7 @@ internal class PatternDetectorTest : Spek({
             val node6 = TestNode()
             val path = listOf(node1, node2, node3, node4, node5, node6)
 
-            val detector = PatternDetector(listOf(path), 1, TestNodeComparator())
+            val detector = FrequentSequenceFinder(listOf(path), 1, TestNodeComparator())
             detector.findFrequentSequences()
 
             assertThat(detector.pathContainsSequence(path, listOf(node5, node6))).isTrue()
@@ -83,7 +83,7 @@ internal class PatternDetectorTest : Spek({
             val node6 = TestNode()
             val path = listOf(node1, node2, node3, node4, node5, node6)
 
-            val detector = PatternDetector(listOf(path), 1, TestNodeComparator())
+            val detector = FrequentSequenceFinder(listOf(path), 1, TestNodeComparator())
             detector.findFrequentSequences()
 
             assertThat(detector.pathContainsSequence(path, listOf(node5, node4))).isFalse()
@@ -96,7 +96,7 @@ internal class PatternDetectorTest : Spek({
             val node5 = TestNode()
             val path = listOf(node2, node3, node4, node5)
 
-            val detector = PatternDetector(listOf(path), 1, TestNodeComparator())
+            val detector = FrequentSequenceFinder(listOf(path), 1, TestNodeComparator())
             detector.findFrequentSequences()
 
             assertThat(detector.pathContainsSequence(path, listOf(node2, node4, node5))).isFalse()
@@ -133,7 +133,7 @@ internal class PatternDetectorTest : Spek({
             val path = listOf(node1, node2, node3)
 
             val paths = listOf(path)
-            val frequent = PatternDetector(paths, 1, TestNodeComparator()).findFrequentSequences()
+            val frequent = FrequentSequenceFinder(paths, 1, TestNodeComparator()).findFrequentSequences()
 
             assertThat(frequent).hasSize(6)
             assertThat(frequent).contains(path)
@@ -156,7 +156,7 @@ internal class PatternDetectorTest : Spek({
             val path3 = listOf(node7, node8, node9, node10)
 
             val paths = listOf(path1, path2, path3)
-            val frequent = PatternDetector(paths, 2, TestNodeComparator()).findFrequentSequences()
+            val frequent = FrequentSequenceFinder(paths, 2, TestNodeComparator()).findFrequentSequences()
 
             assertThat(frequent).isEmpty()
         }
@@ -169,7 +169,7 @@ internal class PatternDetectorTest : Spek({
             val path = listOf(node1, node2, node3, node1, node2, node3)
 
             val paths = listOf(path)
-            val frequent = PatternDetector(paths, 2, TestNodeComparator()).findFrequentSequences()
+            val frequent = FrequentSequenceFinder(paths, 2, TestNodeComparator()).findFrequentSequences()
 
             assertThat(frequent).contains(listOf(node1, node2, node3))
         }
@@ -191,7 +191,7 @@ internal class PatternDetectorTest : Spek({
             val path3 = listOf(node7, node8, node9, node10, node1, node2, node3)
 
             val paths = listOf(path1, path2, path3)
-            val frequent = PatternDetector(paths, 2, TestNodeComparator()).findFrequentSequences()
+            val frequent = FrequentSequenceFinder(paths, 2, TestNodeComparator()).findFrequentSequences()
 
             assertThat(frequent).contains(path1)
         }
@@ -215,7 +215,7 @@ internal class PatternDetectorTest : Spek({
             val path3 = listOf(node7, node8, node9, node10, node1, node2, node3)
 
             val paths = listOf(path1, path2, path3)
-            val patternDetector = PatternDetector(paths, 2, TestNodeComparator())
+            val patternDetector = FrequentSequenceFinder(paths, 2, TestNodeComparator())
 
             patternDetector.findFrequentSequences()
             val patterns = patternDetector.mapFrequentSequencesToPaths()
